+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- concord-server: support for multiple `user` entries in the form's
+`runAs` block;
 - sleep-tasks, concord-server: a way to suspend a process until the
 specified date/time.